Add StatsD arguments and clean up config
Config is now a globvar in the gamq package - no need for playing 'pass
the config'.

The StatsD endpoint metrics are sent to can now be configured via a
command line argument (-statsd) (issue #7). The default behavior is
currently set to look for a statsd broker on localhost:8125 - this should
probably be changed.