Metrics only logged if -statsd flag populated
Metrics will now only be sent to statsd, if the address of a statsd broker
has been provided using the -statsd flag. If this flag is not set -
metrics will still be written to the log, but no attempt will be made to
deliver them to a statsd broker. This closes #7.